# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user: {}
    type: object
  handler.CompatibilityMatchesResponse:
    properties:
      matches:
        items: {}
        type: array
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  handler.LoginRequest:
    properties:
      email:
        example: user@example.com
        format: email
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
...
  handler.RegisterRequest:
    properties:
      email:
        example: user@example.com
        format: email
        type: string
      name:
        example: 홍길동
        type: string
      password:
        example: password123
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  models.Compatibility:
    properties:
      analysis:
        example: 두 사람은 매우 좋은 궁합을 가지고 있습니다.
        type: string
      compatibility_type:
        example: excellent
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      score:
        example: 85.5
        type: number
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user1_id:
        example: 1
        type: integer
      user2_id:
        example: 2
        type: integer
    type: object
...
host: localhost:8080
info:
  title: DoTheFortune API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 이메일과 비밀번호로 로그인합니다. 인증 성공 시 JWT 토큰을 발급하며, 토큰은 응답 본문과 쿠키에 포함됩니다.
        토큰은 24시간 동안 유효합니다.
      parameters:
      - description: 로그인 요청 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 로그인 성공
          schema:
            $ref: '#/definitions/handler.AuthResponse'
        "400":
          description: 잘못된 요청 (이메일 형식 오류 등)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 인증 실패 (이메일 또는 비밀번호 불일치)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 로그인
      tags:
      - auth
...
  /auth/me:
    get:
      consumes:
      - application/json
      description: 현재 인증된 사용자의 ID를 반환합니다. Authorization 헤더의 Bearer 토큰 또는 쿠키의 토큰을 사용하여
        인증합니다.
      produces:
      - application/json
      responses:
        "200":
          description: '사용자 ID 반환"  example:"{\"user_id\": 1}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 인증 실패 (토큰 없음, 만료, 또는 유효하지 않음)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 현재 사용자 정보 조회
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 새로운 사용자를 등록합니다. 이메일, 비밀번호(최소 6자), 이름을 입력받아 계정을 생성하고 JWT 토큰을 발급합니다.
        토큰은 쿠키에도 저장됩니다.
      parameters:
      - description: 회원가입 요청 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 회원가입 성공
          schema:
            $ref: '#/definitions/handler.AuthResponse'
        "400":
          description: 잘못된 요청 (이메일 형식 오류, 비밀번호 길이 부족, 이미 존재하는 이메일 등)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 서버 내부 오류
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 회원가입
      tags:
      - auth
  /compatibility/:
    get:
      consumes:
      - application/json
      description: 현재 사용자와 다른 사용자 간의 저장된 궁합 정보를 조회합니다. 저장된 궁합이 없으면 자동으로 계산하여 반환합니다.
      parameters:
      - description: 상대방 사용자 ID
        in: query
        minimum: 1
        name: user2_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 궁합 정보 조회 성공
          schema:
            $ref: '#/definitions/models.Compatibility'
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 서버 내부 오류
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 궁합 조회
      tags:
      - compatibility
  /compatibility/best:
    get:
      consumes:
      - application/json
      description: 현재 사용자와 가장 좋은 궁합을 가진 사용자들을 궁합 점수 순으로 반환합니다. 점수가 높은 순서대로 정렬되어 반환됩니다.
      parameters:
      - default: 10
        description: 반환할 최대 개수
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 최고 궁합 목록
          schema:
            $ref: '#/definitions/handler.CompatibilityMatchesResponse'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 서버 내부 오류
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 최고 궁합 목록 조회
      tags:
      - compatibility
  /compatibility/calculate:
    get:
      consumes:
      - application/json
      description: 현재 사용자와 다른 사용자 간의 궁합을 계산합니다. 두 사용자의 사주 정보를 비교하여 궁합 점수(0-100)와 분석
        결과를 반환합니다. 계산 결과는 데이터베이스에 저장되며, 이후 조회 시 재계산 없이 저장된 결과를 반환합니다.
      parameters:
      - description: 상대방 사용자 ID
        in: query
        minimum: 1
        name: user2_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 궁합 계산 성공
          schema:
            $ref: '#/definitions/models.Compatibility'
        "400":
          description: 잘못된 요청 (자기 자신과의 궁합 계산 시도 등)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 서버 내부 오류 또는 사주 정보 없음
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 궁합 계산
      tags:
      - compatibility
  /compatibility/worst:
    get:
      consumes:
      - application/json
      description: 현재 사용자와 가장 나쁜 궁합을 가진 사용자들을 궁합 점수 순으로 반환합니다. 점수가 낮은 순서대로 정렬되어 반환됩니다.
      parameters:
      - default: 10
        description: 반환할 최대 개수
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 최악 궁합 목록
          schema:
            $ref: '#/definitions/handler.CompatibilityMatchesResponse'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 서버 내부 오류
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 최악 궁합 목록 조회
      tags:
      - compatibility
...
securityDefinitions:
  BearerAuth:
    description: 'JWT 토큰을 사용한 인증. "Bearer " 뒤에 공백을 두고 JWT 토큰을 입력하세요. 예: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"